+++ packages/libs/common/src/index.ts
@@ -1,4 +1,5 @@
 export * from './utils';
+export * from './transformation';
 export * from './validation';
 export * from './mappings';
 export * from './models';