+++ packages/libs/common/src/validation/index.ts
@@ -1,5 +1,5 @@
 export * from './model.validator';
 export * from './validation.utils';
 export * from './validators.utils';
-export * from './validators';
+export * from './decorators';
 export * from './interfaces/validation-result.interface';